forget password

+++ database/migrations/2021_01_24_214520_add_field_rate_on_table_customer.php
@@ -14,7 +14,7 @@ class AddFieldRateOnTableCustomer extends Migration
     public function up()
     { 
         Schema::table('customers', function (Blueprint $table) {
-            $table->integer('rate')->default(0)->after('avis');
+            $table->decimal('rate')->default(0)->after('avis');
             $table->integer('delivry_count')->default(0)->after('rate');
         });
     }
