+++ database/migrations/2021_01_26_214521_add_field_rate_on_table_adverts.php
@@ -15,7 +15,7 @@ class AddFieldRateOnTableAdverts extends Migration
     {  
          
         Schema::table('adverts', function (Blueprint $table) {
-            $table->integer('rate')->nullable()->default(0);
+            $table->decimal('rate')->nullable()->default(0);
 			$table->text('comment', 65535)->nullable()->after('rate'); 
         }); 
     }
